Copied the relevant test methods from ArtistServiceImplTest to ArtistSearchServiceImplTest, reset all mocks of the test class ArtistsRestControllerTest
