+++ webapp/src/test/java/rocks/metaldetector/service/artist/ArtistServiceImplTest.java
@@ -78,9 +78,6 @@ class ArtistServiceImplTest implements WithAssertions {
   @InjectMocks
   private ArtistServiceImpl underTest;
 
-  @InjectMocks
-  private ArtistSearchServiceImpl underTestSearch;
-
   private ArtistEntity artistEntity;
   private ArtistDto artistDto;
 
@@ -220,255 +217,4 @@ class ArtistServiceImplTest implements WithAssertions {
     verify(artistRepository, times(1)).existsByExternalIdAndSource(EXTERNAL_ID, ARTIST_SOURCE);
   }
 
-  @Nested
-  @DisplayName("Test discogs searching")
-  class DiscogsSearchTest {
-
-    @Test
-    @DisplayName("Should pass provided arguments to discogs service")
-    void searchDiscogsByName_should_pass_arguments() {
-      // given
-      var artistQueryString = "the query";
-      var pageable = PageRequest.of(1, 10);
-      doReturn(UUID.randomUUID().toString()).when(currentPublicUserIdSupplier).get();
-      doReturn(Optional.of(userEntityMock)).when(userRepository).findByPublicId(anyString());
-      doReturn(DiscogsArtistSearchResultDtoFactory.createDefault()).when(discogsService).searchArtistByName(any(), anyInt(), anyInt());
-      doReturn(ArtistSearchResponseFactory.discogs()).when(searchResponseTransformer).transformDiscogs(any());
-
-      // when
-      underTestSearch.searchDiscogsByName(artistQueryString, pageable);
-
-      // then
-      verify(discogsService, times(1)).searchArtistByName(artistQueryString, pageable.getPageNumber(), pageable.getPageSize());
-    }
-
-    @Test
-    @DisplayName("Should call searchResultTransformer with discogs result")
-    void searchDiscogsByName_should_call_result_transformer() {
-      // given
-      var expectedSearchResults = DiscogsArtistSearchResultDtoFactory.createDefault();
-      doReturn(UUID.randomUUID().toString()).when(currentPublicUserIdSupplier).get();
-      doReturn(Optional.of(userEntityMock)).when(userRepository).findByPublicId(anyString());
-      doReturn(expectedSearchResults).when(discogsService).searchArtistByName(any(), anyInt(), anyInt());
-      doReturn(ArtistSearchResponseFactory.discogs()).when(searchResponseTransformer).transformDiscogs(any());
-
-      // when
-      underTestSearch.searchDiscogsByName("the query", PageRequest.of(1, 10));
-
-      // then
-      verify(searchResponseTransformer, times(1)).transformDiscogs(expectedSearchResults);
-    }
-
-    @Test
-    @DisplayName("Should return searchResultTransformer's result")
-    void searchDiscogsByName_should_return_results() {
-      // given
-      var expectedSearchResult = ArtistSearchResponseFactory.discogs();
-      doReturn(UUID.randomUUID().toString()).when(currentPublicUserIdSupplier).get();
-      doReturn(Optional.of(userEntityMock)).when(userRepository).findByPublicId(anyString());
-      doReturn(DiscogsArtistSearchResultDtoFactory.createDefault()).when(discogsService).searchArtistByName(any(), anyInt(), anyInt());
-      doReturn(expectedSearchResult).when(searchResponseTransformer).transformDiscogs(any());
-
-      // when
-      ArtistSearchResponse result = underTestSearch.searchDiscogsByName("the query", PageRequest.of(1, 10));
-
-      // then
-      assertThat(result).isEqualTo(expectedSearchResult);
-    }
-
-    @Test
-    @DisplayName("Should mark all already followed artists")
-    void searchDiscogsByName_should_mark_already_followed_artists() {
-      // given
-      var discogssearchresults = ArtistSearchResponseFactory.discogs();
-      discogssearchresults.setSearchResults(createListOfSearchResultEntries(List.of("1", "2", "3")));
-      doReturn(UUID.randomUUID().toString()).when(currentPublicUserIdSupplier).get();
-      doReturn(Optional.of(userEntityMock)).when(userRepository).findByPublicId(anyString());
-      when(userEntityMock.isFollowing(anyString())).then(invocationOnMock -> {
-        String artistId = invocationOnMock.getArgument(0);
-        return artistId.equals("1") || artistId.equals("3");
-      });
-      doReturn(discogssearchresults).when(searchResponseTransformer).transformDiscogs(any());
-
-      // when
-      var searchResults = underTestSearch.searchDiscogsByName("the query", PageRequest.of(1, 10));
-
-      // then
-      assertThat(searchResults.getSearchResults().get(0).isFollowed()).isTrue();
-      assertThat(searchResults.getSearchResults().get(1).isFollowed()).isFalse();
-      assertThat(searchResults.getSearchResults().get(2).isFollowed()).isTrue();
-    }
-
-    @Test
-    @DisplayName("Should get current user")
-    void searchDiscogsByName_should_get_user() {
-      // given
-      var artistQueryString = "the query";
-      var pageable = PageRequest.of(1, 10);
-      var userId = "userId";
-      doReturn(userId).when(currentPublicUserIdSupplier).get();
-      doReturn(Optional.of(userEntityMock)).when(userRepository).findByPublicId(anyString());
-      doReturn(DiscogsArtistSearchResultDtoFactory.createDefault()).when(discogsService).searchArtistByName(any(), anyInt(), anyInt());
-      doReturn(ArtistSearchResponseFactory.discogs()).when(searchResponseTransformer).transformDiscogs(any());
-
-      // when
-      underTestSearch.searchDiscogsByName(artistQueryString, pageable);
-
-      // then
-      verify(currentPublicUserIdSupplier, times(1)).get();
-      verify(userRepository, times(1)).findByPublicId(userId);
-    }
-
-    @Test
-    @DisplayName("Should throw exception when user not found")
-    void searchDiscogsByName_should_throw_exception() {
-      // given
-      var artistQueryString = "the query";
-      var pageable = PageRequest.of(1, 10);
-      var userId = "userId";
-      doReturn(userId).when(currentPublicUserIdSupplier).get();
-      doThrow(new ResourceNotFoundException(userId)).when(userRepository).findByPublicId(anyString());
-
-      // when
-      var throwable = catchThrowable(() -> underTestSearch.searchDiscogsByName(artistQueryString, pageable));
-
-      // then
-      assertThat(throwable).isInstanceOf(ResourceNotFoundException.class);
-      assertThat(throwable).hasMessageContaining(userId);
-    }
-
-    private List<ArtistSearchResponseEntryDto> createListOfSearchResultEntries(List<String> externalIds) {
-      return List.of(
-          ArtistSearchResponseEntryDtoFactory.withId(externalIds.get(0)),
-          ArtistSearchResponseEntryDtoFactory.withId(externalIds.get(1)),
-          ArtistSearchResponseEntryDtoFactory.withId(externalIds.get(2))
-      );
-    }
-  }
-
-  @Nested
-  @DisplayName("Test spotify searching")
-  class SpotifySearchTest {
-
-    @Test
-    @DisplayName("Should pass provided arguments to spotify service")
-    void searchSpotifyByName_should_pass_arguments() {
-      // given
-      var artistQueryString = "the query";
-      var pageable = PageRequest.of(1, 10);
-      doReturn(UUID.randomUUID().toString()).when(currentPublicUserIdSupplier).get();
-      doReturn(Optional.of(userEntityMock)).when(userRepository).findByPublicId(anyString());
-      doReturn(SpotifyArtistSearchResultDtoFactory.createDefault()).when(spotifyService).searchArtistByName(any(), anyInt(), anyInt());
-      doReturn(ArtistSearchResponseFactory.spotify()).when(searchResponseTransformer).transformSpotify(any());
-
-      // when
-      underTestSearch.searchSpotifyByName(artistQueryString, pageable);
-
-      // then
-      verify(spotifyService, times(1)).searchArtistByName(artistQueryString, pageable.getPageNumber(), pageable.getPageSize());
-    }
-
-    @Test
-    @DisplayName("Should call searchResultTransformer with spotify result")
-    void searchSpotifyByName_should_call_result_transformer() {
-      // given
-      var expectedSearchResults = SpotifyArtistSearchResultDtoFactory.createDefault();
-      doReturn(UUID.randomUUID().toString()).when(currentPublicUserIdSupplier).get();
-      doReturn(Optional.of(userEntityMock)).when(userRepository).findByPublicId(anyString());
-      doReturn(expectedSearchResults).when(spotifyService).searchArtistByName(any(), anyInt(), anyInt());
-      doReturn(ArtistSearchResponseFactory.spotify()).when(searchResponseTransformer).transformSpotify(any());
-
-      // when
-      underTestSearch.searchSpotifyByName("the query", PageRequest.of(1, 10));
-
-      // then
-      verify(searchResponseTransformer, times(1)).transformSpotify(expectedSearchResults);
-    }
-
-    @Test
-    @DisplayName("Should return searchResultTransformer's result")
-    void searchSpotifyByName_should_return_results() {
-      // given
-      var expectedSearchResult = ArtistSearchResponseFactory.spotify();
-      doReturn(UUID.randomUUID().toString()).when(currentPublicUserIdSupplier).get();
-      doReturn(Optional.of(userEntityMock)).when(userRepository).findByPublicId(anyString());
-      doReturn(SpotifyArtistSearchResultDtoFactory.createDefault()).when(spotifyService).searchArtistByName(any(), anyInt(), anyInt());
-      doReturn(expectedSearchResult).when(searchResponseTransformer).transformSpotify(any());
-
-      // when
-      ArtistSearchResponse result = underTestSearch.searchSpotifyByName("the query", PageRequest.of(1, 10));
-
-      // then
-      assertThat(result).isEqualTo(expectedSearchResult);
-    }
-
-    @Test
-    @DisplayName("Should mark all already followed artists")
-    void searchSpotifyByName_should_mark_already_followed_artists() {
-      // given
-      var spotifySearchResults = ArtistSearchResponseFactory.spotify();
-      spotifySearchResults.setSearchResults(createListOfSearchResultEntries(List.of("1", "2", "3")));
-      doReturn(UUID.randomUUID().toString()).when(currentPublicUserIdSupplier).get();
-      doReturn(Optional.of(userEntityMock)).when(userRepository).findByPublicId(anyString());
-      when(userEntityMock.isFollowing(anyString())).then(invocationOnMock -> {
-        String artistId = invocationOnMock.getArgument(0);
-        return artistId.equals("1") || artistId.equals("3");
-      });
-      doReturn(spotifySearchResults).when(searchResponseTransformer).transformSpotify(any());
-
-      // when
-      var searchResults = underTestSearch.searchSpotifyByName("the query", PageRequest.of(1, 10));
-
-      // then
-      assertThat(searchResults.getSearchResults().get(0).isFollowed()).isTrue();
-      assertThat(searchResults.getSearchResults().get(1).isFollowed()).isFalse();
-      assertThat(searchResults.getSearchResults().get(2).isFollowed()).isTrue();
-    }
-
-    @Test
-    @DisplayName("Should get current user")
-    void searchSpotifyByName_should_get_user() {
-      // given
-      var artistQueryString = "the query";
-      var pageable = PageRequest.of(1, 10);
-      var userId = "userId";
-      doReturn(userId).when(currentPublicUserIdSupplier).get();
-      doReturn(Optional.of(userEntityMock)).when(userRepository).findByPublicId(anyString());
-      doReturn(SpotifyArtistSearchResultDtoFactory.createDefault()).when(spotifyService).searchArtistByName(any(), anyInt(), anyInt());
-      doReturn(ArtistSearchResponseFactory.spotify()).when(searchResponseTransformer).transformSpotify(any());
-
-      // when
-      underTestSearch.searchSpotifyByName(artistQueryString, pageable);
-
-      // then
-      verify(currentPublicUserIdSupplier, times(1)).get();
-      verify(userRepository, times(1)).findByPublicId(userId);
-    }
-
-    @Test
-    @DisplayName("Should throw exception when user not found")
-    void searchSpotifyByName_should_throw_exception() {
-      // given
-      var artistQueryString = "the query";
-      var pageable = PageRequest.of(1, 10);
-      var userId = "userId";
-      doReturn(userId).when(currentPublicUserIdSupplier).get();
-      doThrow(new ResourceNotFoundException(userId)).when(userRepository).findByPublicId(anyString());
-
-      // when
-      var throwable = catchThrowable(() -> underTestSearch.searchSpotifyByName(artistQueryString, pageable));
-
-      // then
-      assertThat(throwable).isInstanceOf(ResourceNotFoundException.class);
-      assertThat(throwable).hasMessageContaining(userId);
-    }
-
-    private List<ArtistSearchResponseEntryDto> createListOfSearchResultEntries(List<String> externalIds) {
-      return List.of(
-          ArtistSearchResponseEntryDtoFactory.withId(externalIds.get(0)),
-          ArtistSearchResponseEntryDtoFactory.withId(externalIds.get(1)),
-          ArtistSearchResponseEntryDtoFactory.withId(externalIds.get(2))
-      );
-    }
-  }
 }
