+++ webapp/src/test/java/rocks/metaldetector/web/controller/rest/ArtistsRestControllerTest.java
@@ -77,7 +77,7 @@ class ArtistsRestControllerTest implements WithAssertions {
 
     @AfterEach
     void tearDown() {
-      reset(artistSearchService);
+      reset(artistService, artistSearchService, followArtistService);
     }
 
     @Test
